Refactor: migrate pagination to database package
